Fix for Issue #23. Made sure log backups had the correct year. Also made sure restarting the flight computer wouldn't overwrite the log. -splice can't write to files in append mode -use lseek64 to get the right offset for the splice operation -change file pointers to file descriptors to allow lseek64() to work properly -try doing positive offset

+++ MOSES_FC/logger.c
@@ -98,7 +98,7 @@ void copy_log_to_disk() {
 
     /* Move the cursor to the end of the backup log*/
     off64_t out_offset = lseek64(out_fd, 0, SEEK_END);
-    printf("Out offset is: %d\n", (int) out_offset);
+    printf("Out offset is: %ld\n", (long int) out_offset);
     if (out_offset == -1)
         printf("%d - %s\n", errno, strerror(errno));
